+++ Sourcery/Models/Method.swift
@@ -23,16 +23,18 @@ final class MethodParameter: NSObject, SourceryModel, Typed {
     // sourcery: skipEquality, skipDescription, skipCoding
     internal var __parserData: Any?
 
-    init(argumentLabel: String?, name: String = "", typeName: TypeName) {
+    init(argumentLabel: String?, name: String = "", typeName: TypeName, type: Type? = nil) {
         self.typeName = typeName
         self.argumentLabel = argumentLabel
         self.name = name
+        self.type = type
     }
 
-    init(name: String = "", typeName: TypeName) {
+    init(name: String = "", typeName: TypeName, type: Type? = nil) {
         self.typeName = typeName
         self.argumentLabel = name
         self.name = name
+        self.type = type
     }
 
     // sourcery:inline:MethodParameter.AutoCoding
@@ -71,6 +73,10 @@ final class Method: NSObject, SourceryModel, Annotated {
 
     /// Name of the return type
     var returnTypeName: TypeName
+
+    var actualReturnTypeName: TypeName {
+        return returnTypeName.actualTypeName ?? returnTypeName
+    }
 
     /// Actual method return type, if known.
     // sourcery: skipEquality